GCE node templates refactoring
