+++ cluster-autoscaler/cloudprovider/gce/templates_test.go
